// apps/web/src/lib/media/types/media.types.ts
export const MediaType = {
  AUDIO: 'audio',
  IMAGE: 'image',
  BANNER: 'banner',
  PDF: 'pdf',
  VIDEO: 'video',
  DOCUMENT: 'document',
} as const;
export type MediaType = (typeof MediaType)[keyof typeof MediaType];